+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support to download Sentinel-2 products from Google Cloud storage.
+- Added support to download by text match.
+- Added `download by-geometry` and `download by-text` to Fire commands.
 
 
 ## Released